+++ packages/browser-wallet/src/popup/store/token.ts
@@ -2,10 +2,8 @@
 import { Atom, atom } from 'jotai';
 import { mapRecordValues } from 'wallet-common-helpers/src/utils/basicHelpers';
 import { atomFamily } from 'jotai/utils';
-import { Buffer } from 'buffer/';
-import { AccountAddress, JsonRpcClient } from '@concordium/web-sdk';
-import uleb128 from 'leb128/unsigned';
 import { ChromeStorageKey, TokenIdAndMetadata, TokenMetadata, TokenStorage } from '@shared/storage/types';
+import { getTokenBalance } from '@shared/utils/token-helpers';
 import { AsyncWrapper, atomWithChromeStorage } from './utils';
 import { jsonRpcClientAtom } from './settings';
 
@@ -41,60 +39,6 @@ export const tokensAtom = atom<AsyncWrapper<Record<string, Record<string, TokenI
     };
 });
 
-/**
- * Serialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#id3
- */
-const serializeBalanceParameter = (tokenIndex: string, accountAddress: string) => {
-    const queries = Buffer.alloc(2);
-    queries.writeUInt16LE(1, 0); // 1 query
-
-    const token = Buffer.from(tokenIndex, 'hex');
-    const tokenLength = Buffer.alloc(1);
-    tokenLength.writeUInt8(token.length, 0);
-
-    const addressType = Buffer.alloc(1); // Account address type
-    const address = new AccountAddress(accountAddress).decodedAddress;
-
-    return Buffer.concat([queries, tokenLength, token, addressType, address]);
-};
-
-/**
- * Deserialized based on cis-2 documentation: https://proposals.concordium.software/CIS/cis-2.html#response
- */
-const deserializeBalanceAmount = (value: string): bigint => {
-    const buf = Buffer.from(value, 'hex');
-    const amount = uleb128.decode(buf.slice(2)); // ignore first 2 bytes as we only expect 1 tokenamount
-
-    return BigInt(amount);
-};
-
-const getBalance = async (
-    client: JsonRpcClient,
-    contractIndex: string,
-    tokenIndex: string,
-    accountAddress: string
-): Promise<bigint> => {
-    const index = BigInt(contractIndex);
-    const subindex = 0n;
-    const instanceInfo = await client.getInstanceInfo({ index, subindex });
-
-    if (instanceInfo === undefined) {
-        return 0n;
-    }
-
-    const result = await client.invokeContract({
-        contract: { index, subindex },
-        method: `${instanceInfo.name.substring(5)}.balanceOf`,
-        parameter: serializeBalanceParameter(tokenIndex, accountAddress),
-    });
-
-    if (result === undefined || result.tag === 'failure' || result.returnValue === undefined) {
-        return 0n;
-    }
-
-    return deserializeBalanceAmount(result.returnValue);
-};
-
 const tbf = atomFamily<string, Atom<Promise<bigint>>>((tokenId: string) => {
     const parts = tokenId.split('.');
 
@@ -106,7 +50,7 @@ const tbf = atomFamily<string, Atom<Promise<bigint>>>((tokenId: string) => {
 
     return atom<Promise<bigint>>((get) => {
         const client = get(jsonRpcClientAtom);
-        return getBalance(client, contractIndex, tokenIndex, accountAddress);
+        return getTokenBalance(client, contractIndex, tokenIndex, accountAddress);
     });
 });
