+++ packages/browser-wallet/src/types/leb128.d.ts
@@ -1,5 +1,3 @@
-// import type Buffer from 'buffer/';
-
 declare module 'leb128/unsigned' {
     export function decode(buffer: import('buffer/').Buffer): string;
 }
